Enhance OIDC authentication flow by introducing a new `oidcConfig` function, updating cookie management, and refining the login process. Removed unused constants and improved error handling in the authentication API. Updated related tests to reflect these changes.
