+++ packages/server/src/apis/auth.ts
@@ -1,15 +1,10 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
-import { LOGIN_ERROR_PATH } from '../auth/constants';
-import {
-  clearIdTokenCookie,
-  clearOAuthStateCookie,
-  readOAuthStateCookie,
-  setIdTokenCookie,
-  setOAuthStateCookie,
-} from '../auth/cookies';
-import { buildLoginAuthorization, exchangeAuthorizationCode, oidcConfig, safeReturnTo } from '../auth/oidc';
-import configuration from '../config';
+import { clearAuthCookie, ID_TOKEN_COOKIE, OAUTH_STATE_COOKIE, readOAuthStateCookie } from '../auth/cookies';
+import { buildLoginAuthorization, exchangeAuthorizationCode, safeReturnTo } from '../auth/oidc';
+import { oidcConfig } from '../config';
 import { authLoginRoute, authLogoutRoute, oAuthCallbackRoute } from '../routes/authRoutes';
+
+const LOGIN_ERROR_PATH = '/?error=login_failed';
 
 export function createAuthRouter() {
   const router = new OpenAPIHono();
@@ -21,11 +16,10 @@ export function createAuthRouter() {
     }
 
     try {
-      const { authorizationUrl, state, codeVerifier } = await buildLoginAuthorization();
-      setOAuthStateCookie(c, {
-        state,
-        code_verifier: codeVerifier,
-        return_to: safeReturnTo(c.req.valid('query').return_to),
+      const authorizationUrl = await buildLoginAuthorization({
+        context: c,
+        oidc,
+        returnTo: c.req.valid('query').return_to,
       });
       return c.redirect(authorizationUrl, 302);
     } catch {
@@ -41,7 +35,7 @@ export function createAuthRouter() {
 
     const query = c.req.valid('query');
     const pending = readOAuthStateCookie(c);
-    clearOAuthStateCookie(c);
+    clearAuthCookie({ context: c, name: OAUTH_STATE_COOKIE });
 
     if (!pending) {
       return c.redirect(LOGIN_ERROR_PATH, 302);
@@ -54,15 +48,13 @@ export function createAuthRouter() {
     }
 
     try {
-      const requestUrl = new URL(c.req.url);
-      const callbackUrl = new URL(requestUrl.pathname + requestUrl.search, configuration.PUBLIC_BASE_URL);
-      const { idToken, expiresAtSeconds } = await exchangeAuthorizationCode({
-        callbackUrl,
+      await exchangeAuthorizationCode({
+        context: c,
+        oidc,
+        code: query.code,
         codeVerifier: pending.code_verifier,
         expectedState: pending.state,
       });
-      const maxAge = Math.max(0, expiresAtSeconds - Math.floor(Date.now() / 1000));
-      setIdTokenCookie(c, idToken, maxAge);
       return c.redirect(safeReturnTo(pending.return_to), 302);
     } catch {
       return c.redirect(LOGIN_ERROR_PATH, 302);
@@ -70,7 +62,8 @@ export function createAuthRouter() {
   });
 
   router.openapi(authLogoutRoute, c => {
-    clearIdTokenCookie(c);
+    // Cookie deletion is idempotent, so logout also succeeds when no cookie exists.
+    clearAuthCookie({ context: c, name: ID_TOKEN_COOKIE });
     return c.body(null, 204);
   });
