--- packages/server/src/auth/constants.ts
@@ -1,6 +1,0 @@
-export const OAUTH_STATE_COOKIE = 'oauth_state';
-export const ID_TOKEN_COOKIE = 'id_token';
-export const CALLBACK_PATH = '/api/v1/auth/callback';
-export const LOGIN_ERROR_PATH = '/?error=login_failed';
-export const OAUTH_STATE_MAX_AGE_SECONDS = 10 * 60;
-export const OIDC_SCOPES = 'openid email profile';
