+++ packages/server/src/auth/oidc.ts
@@ -1,39 +1,41 @@
+import type { Context } from 'hono';
 import * as client from 'openid-client';
-import configuration, { type OIDCConfig } from '../config';
-import { CALLBACK_PATH, OIDC_SCOPES } from './constants';
+import type { Logger } from 'winston';
+import configuration, { type OIDCConfig, oidcConfig } from '../config';
+import { ID_TOKEN_COOKIE, OAUTH_STATE_COOKIE, setAuthCookie } from './cookies';
 
-let oidcConfiguration: client.Configuration | undefined;
+const CALLBACK_PATH = '/api/v1/auth/callback';
+const OAUTH_STATE_MAX_AGE_SECONDS = 10 * 60;
+const OIDC_SCOPES = 'openid email profile';
 
-export function oidcConfig(): OIDCConfig | undefined {
-  if (configuration.STANDALONE) {
-    return undefined;
-  }
-  return configuration.OIDC;
-}
+let initializedOidc: { oidc: OIDCConfig; client: client.Configuration } | undefined;
 
 /** Discover the IdP once at process startup when OIDC is configured. */
-export async function initOidc(): Promise<void> {
+export async function initOidc(options: { logger: Logger }): Promise<void> {
   const oidc = oidcConfig();
   if (!oidc) {
+    options.logger.warn('OIDC is not configured; browser login is disabled');
     return;
   }
 
+  options.logger.info('OIDC is configured', { issuer: oidc.OIDC_ISSUER_URL });
   const issuer = new URL(oidc.OIDC_ISSUER_URL);
   const discoveryUrl = new URL(
     `${issuer.origin}${issuer.pathname.replace(/\/$/, '')}/.well-known/openid-configuration`,
   );
   discoveryUrl.searchParams.set('client_id', oidc.OIDC_CLIENT_ID);
-  oidcConfiguration = await client.discovery(discoveryUrl, oidc.OIDC_CLIENT_ID, oidc.OIDC_CLIENT_SECRET);
+  const discoveredClient = await client.discovery(discoveryUrl, oidc.OIDC_CLIENT_ID, oidc.OIDC_CLIENT_SECRET);
+  initializedOidc = { oidc, client: discoveredClient };
 }
 
-function requireOidcConfiguration(): client.Configuration {
-  if (!oidcConfiguration) {
+function requireOidcConfiguration(oidc: OIDCConfig): client.Configuration {
+  if (initializedOidc?.oidc !== oidc) {
     throw new Error('OIDC was not initialized; call initOidc() during server startup');
   }
-  return oidcConfiguration;
+  return initializedOidc.client;
 }
 
-export function authCallbackUrl(): string {
+function authCallbackUrl(): string {
   return `${configuration.PUBLIC_BASE_URL}${CALLBACK_PATH}`;
 }
 
@@ -52,40 +54,58 @@ export function safeReturnTo(value: string | undefined): string {
   return '/';
 }
 
-export async function buildLoginAuthorization(): Promise<{
-  authorizationUrl: string;
-  state: string;
-  codeVerifier: string;
-}> {
+export async function buildLoginAuthorization(options: {
+  context: Context;
+  oidc: OIDCConfig;
+  returnTo: string | undefined;
+}): Promise<string> {
+  const returnTo = safeReturnTo(options.returnTo);
   const codeVerifier = client.randomPKCECodeVerifier();
   const state = client.randomState();
-  const authorizationUrl = client.buildAuthorizationUrl(requireOidcConfiguration(), {
+  const authorizationUrl = client.buildAuthorizationUrl(requireOidcConfiguration(options.oidc), {
     redirect_uri: authCallbackUrl(),
     scope: OIDC_SCOPES,
     code_challenge: await client.calculatePKCECodeChallenge(codeVerifier),
     code_challenge_method: 'S256',
     state,
   });
-  return {
-    authorizationUrl: authorizationUrl.href,
-    state,
-    codeVerifier,
-  };
+  setAuthCookie({
+    context: options.context,
+    name: OAUTH_STATE_COOKIE,
+    value: JSON.stringify({
+      state,
+      code_verifier: codeVerifier,
+      return_to: returnTo,
+    }),
+    maxAgeSeconds: OAUTH_STATE_MAX_AGE_SECONDS,
+  });
+  return authorizationUrl.href;
 }
 
 export async function exchangeAuthorizationCode(options: {
-  callbackUrl: URL;
+  context: Context;
+  oidc: OIDCConfig;
+  code: string;
   codeVerifier: string;
   expectedState: string;
-}): Promise<{ idToken: string; expiresAtSeconds: number }> {
-  const tokens = await client.authorizationCodeGrant(requireOidcConfiguration(), options.callbackUrl, {
+}): Promise<void> {
+  // redirect_uri is fixed; openid-client still needs the IdP response query (code + state).
+  const callbackUrl = new URL(authCallbackUrl());
+  callbackUrl.searchParams.set('code', options.code);
+  callbackUrl.searchParams.set('state', options.expectedState);
+  const tokens = await client.authorizationCodeGrant(requireOidcConfiguration(options.oidc), callbackUrl, {
     pkceCodeVerifier: options.codeVerifier,
     expectedState: options.expectedState,
   });
   const idToken = tokens.id_token;
-  const claims = tokens.claims();
-  if (!idToken || !claims?.exp) {
+  const expiresAtSeconds = tokens.claims()?.exp;
+  if (!idToken || !expiresAtSeconds) {
     throw new Error('Token response missing id_token or exp');
   }
-  return { idToken, expiresAtSeconds: claims.exp };
+  setAuthCookie({
+    context: options.context,
+    name: ID_TOKEN_COOKIE,
+    value: idToken,
+    maxAgeSeconds: Math.max(0, expiresAtSeconds - Math.floor(Date.now() / 1000)),
+  });
 }
